requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.23.2
python-telegram-bot==20.3
gspread==5.8.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.30.0
pydantic-settings==2.3.0
pytest==7.0.0